Loop control and empty statment now supported
Loop control now known now includes:
 * redo
 * next
 * break

the `redo` statement does jump to the beginning of the loop content,
without executing the condition again, while the `next` will do execute
the loop conditon again and then continue to execute its content.

break, obviously, like other languages, break from a loop.

and additionally the `continue` is the alias of `next`,
and the `last` statement is the alias of `break`
 (I am a Ruby & Perl & C programmer :-D )

Note now the loop is only provided by `while` statement, so they will
just be able to use inside while.

BUG:

If there is a function call inside a loop, and the call has no loop
structure but a loop control statement is given, it will not just report
error as expected, rather, the control will affect the outer level of loop.