Debugging system is now completely rewritten!
In the code, now it could use `EMIT' keyword for emitting debugging
values, and the receiver would arrange them into a array.

All of this are based on a variable controled, it is `$CS_DEBUG', if it
is set(not nil), the emission will be turned on, otherwise, for example,
in the release version, it will be not used anymore.

Accompany the changes of debugging methods, it could adapt to unit test
now also, so I wrote some code for easily testing, the code is put
at ./test/common.rb, which includes some specific assertions, and some
convenient tools to test. Those, currently are:
 - simple_test
 - simple_test_raise

The first one check equality of expected emissions and the actual
emissions, and the second one test if it catches specific error.

Both of the function accept the expected stuff as the argument, and a
block that returns code string. The functions will parse and execute the
code inside and assert the result if it is matched.

Ain't these really big changes?

It's really late & I gotta go to bed now, nice dream :)