Support for logic operators
They are:
    - !
    - &&
    - ||

The logic negative operator(!) takes a expression. If the result of
evaluation of the expression is false, return true, otherwise, return false.

Then logic and operator(&&) will comfirm both side of the operator is
true, and return true if that is, otherwise, return false. Note: this
operator will return false if the first operand is false and do not
evaluate the second one any more. (a.k.a. short circuit)

And the logic or operator(||) will return the operands' value if any of
them are true, as it is a short circuit too, which means, if the first
operand is known as true, it will return the first operand without
evaluating the second one, otherwise, the second one will get evaluated.
If both two operands is not evaluated as true, the final result will be false.