It is a really great reconstruction
I used three days to reconstruct the executor part, with much more
'rubish' code and some functional improvement.

The basic purpose is just provide a manner to run from JSON serialized
syntax tree. But it not just doing that now.

Now it is much better structured, and it looks more elegant in the code.

However, it is not fully completed yet, as one of most important
component is not added into yet. That is a class to store variables,
which is temporarily named as `SymbolTable`.

Good luck & I'll finish it tomorrow.