Finally, it runs.
I spend whole night to do improvement and error fixing.

In the morning, the first test code runs finally and thus I saw the
correct result delightly.

    EMIT 1+1;

and it gives me `[2]' expectedly.

Most errors fixed are syntatic, and most of them are just coursed by
mistyping names, the other part are uses inproperly.

It is difficult to explain everything well, as you can check them all
out from the `diff's.

Be confident on your four nights' work, it is really amazing. The result
already prooved that your basic design is flexible.

Anyway, the reconstruction could be told as terminated. Of course, it is
not fully stabled, and with many hidden bugs, but them all, including
other function extensions and bug fixing, are just the postwords.

yet another night.

GOOD LUCK & HAVE A NICE DREAM.