Repassed all the tests
By modified a batch of source files, I, finally make the code to pass
the tests. It deserve to be congratulated very much.

As you see, most failure are still the code's problem, except of one,
and I'll talk something about it below.

Read the code:

> local_var = 1;
> def foo() {
>     ...
> }
>
> def bar() {
>     foo();
> }
> def baz() {
>     func_local_var = 1;
>     bar();
> }
>
> bar();

Originally I store a function as a variable in the CURRENT name scope,
that means, `foo` and `local_var` should belong in the same scope in
the program. But obviously a function calling operation should create a
completely new scope to execute its code and prevent to pollution the
parent's, which means, in the function `bar`, it will not be able to
access variable `local_var` outside, just like that it cannot access the
`func_local_var` in baz. So that's it.

And you know, function calling has these steps:
1. Find a variable with the function's name
2. Process & evaluate arguments
3. Create a sole CallStack, and put the function & arguments into it
4. Execute the CallStack

In the first step, it will search current and parent scope for the
function, while if currently it is in a calling scope, then it will be
isolated and can only access current scope and the global scope that
defined in a common Runtime. But a function's definition is in so-called
"local scope", so, the code in the above demonstration cannot normally
access to function `foo` in the calling of the function `bar`.

Not strange, it is a fault on design. I don't know why I didn't consider
about this when I designing runstack and callstack.

Hence, I just found a compromising solution. That is, the definition of
a function will be as a global variable by default.

That's a flexible solution as far, but I have to consider about how
global scope should be regarded in a multiple file project and how
things like anonymous functions(a.k.a. lambda) should be act as.

That's it. I feel much better now.