Completely rewritten the residual scanner module
You know few days before I almost rewritten the whole program in a more
elegant and friendly form. Today, I seems finished the last residual
part, the Scanner.

Frankly, the reason why I didn't rewrite this in those days is that the
lexer always works well, it is tending to steady.

You know, this is the first part I've finished in this version of
cscript. At that time, near two or three months ago, I even can't handle
Ruby as practiced as now. I even struggled in the dynamic binding and
can't get a solution for that.

But now, more and more I am tending to write program that would be
looking perfect, easy to read(programmer friendly), and efficient. As a
perfectionist, of course looking elegant is the most important thing of
code, and usually, in passing, beautiful code is easier to read and
debug.

Changes:
    * Yacc(old Parser), Preprocessor, Scanner are now equal! They as the
      same status in the Parser class.
    * Now the Parser class is compeletely rewritten. It now only
      a few public interfaces from its members(yacc, pp, scanner).
    * As now Preprocessor and Scanner are equal, they could share the
      same Parser [parent], which means, they could now access each
      other conveniently. So they do not need to pass theirselves as
      arguments to commnicate each other. See `Preprocessor#preprocess`
    * `:DROP` action is preprocessor is useless, originally. It could
      be replaced completely by `:PASS`, as they, on behavior, has no
      differences actually.
    * In the handlers of Evaluator, `next` is used to replace `return`
      to return a value and jump directly from the block. The most
      obvious advantage is, I don't ever meed to generate methods for
      each blocks, and could now call them more conveniently.
    * Scanner is now a class instead of a module before. It is almost
      completely rewritten as I used many advanced trick to make it look
      comfortable, look the rule definition's part, cool isn't it.
      And this is the main archivement tonight.
    * Test ultitlies now additionally support argument `cat` and
      `lexer`. The former one do print the source file of the test
      prorgram(no only cscript). And the latter one is used to see the
      lexical parsing result.
    * A small bug fix in a test file.
    * A convenient tool which is on purpose to do all test in script. It
      scan the test files inside the directory and do the tests, and
      report whether it passes the test successfully or not.

New Functions (in the Ultilities module):
    * `proc_to_lambda`, a mata programming method which is used to
      convert a proc into a lambda. (now it is not confirmed work, don't
      use yet)
    * `generate_method`, yet another metaprogramming method to generate
      a unbound method from a given block.
    * `binding_call`, call a block with specific `binding`. It works
      and it is very cool.

Bug Fix:
    * preprocessor macro will not now be executed inside comments.