static scope is now supported
The `static` keyword in cscript is just similar as the same in C.

Same as `global`, `static` do both promiting and declaring. For relative
usage and detailed explaintion, see the commit log of the `global`
feature.

This time I added a special scope(implement as runstack) in a runtime.
However, this scope is not changeless in one program -- it will change
when ever new codes are executed by the runtime.

The new `static scope` as the runstack type of `:static`, it should be
indicated correctly as it is used at name finding in the symbol table
class.

Nothing special.

And another change is now defaultly a function would not be defined in
global scope, rather, static scope instead.

Bug found & fixes:
    * new execution will override current running runstack in the old
      versions, and fixed now, by using a stack to save current ones
      before creating new ones.
    * a global promoting might not success as the old variable will be
      still there in the old versions, and now it is fixed.