Unamed expression return lambda is callable now
The old rule was:

 func_call:  name '(' ... ')'

And the new one is changed to:

 func_call: expr '(' ... ')'

So that the called object could be changed to any expression that
returns a function or a lambda.

A bug is fixed:
    * Create a multiple nested lambda will block `make_binding`
      operation.

No other changes.

I was not very happy for these as the implment of spaces at eol is too
annoying.

Nice dream, maybe.