Ignorable semicolon(;) in statement lists end
Simply, for example, in the former code you can only write as:

    EMIT (-> { "good morning"; })();

Obviously the semicolon in the lambda expression's body is ugly
and redundant. This is what I have improved in this version.

Now you can ignore the semicolon, and with it is okay, too.

And for any blocks, this feature is always usable, for example:

    if (i == 1) { flag = 1; return 0 }

is now okay. But note that the semicolon can only be ignored when
it is in the last place of the last statement of a block.

And this effects on all inline statements, these are currently
supported operations:
    * basic expressions
    * `return` statements
    * loop controls (i.e. redo, next and break)
    * emit macros
    * global/static variable declarations

Enjoy. As I have a long time left from this project. But I am happy
to get back with cscript, and I'll maintain it for longer time.